last brick of pattern test cases for a Alphabet
